add python script for generating update files. - Move code from tools/update_database/follow_db_changes.py to xml_converter.py, so it can be reused. - Add generate_db.py which only does the generate part of follow_db_changes.py
