+++ tools/update_database/follow_db_changes.py
@@ -20,13 +20,15 @@ Note that this script also creates a database with version 0.  This may be
 downloaded manually by people who use Lensfun <= 0.2.8.
 """
 
-import glob, os, subprocess, calendar, json, time, tarfile, io, argparse, shutil, configparser, smtplib, textwrap
+import glob, os, subprocess, json, argparse, shutil, configparser, smtplib, textwrap
 from subprocess import DEVNULL
 from email.mime.text import MIMEText
-from lxml import etree
+
 from github import Github
 from calibration_webserver import owncloud
 
+import xml_converter
+from xml_converter import XMLFile
 
 parser = argparse.ArgumentParser(description="Generate tar balls of the Lensfun database, also for older versions.")
 parser.add_argument("output_path", help="Directory where to put the XML files.  They are put in the db/ subdirectory.  "
@@ -45,39 +47,6 @@ unsuccessful_label = lensfun.get_label("unsuccessful")
 unsuccessful_no_mail_label = lensfun.get_label("unsuccessful-no-mail")
 admin = "{} <{}>".format(config["General"]["admin_name"], config["General"]["admin_email"])
 root = "/tmp/"
-
-
-class XMLFile:
-
-    def __init__(self, root, filepath):
-        self.filepath = filepath
-        self.tree = etree.parse(os.path.join(root, filepath))
-
-    @staticmethod
-    def indent(tree, level=0):
-        i = "\n" + level*"    "
-        if len(tree):
-            if not tree.text or not tree.text.strip():
-                tree.text = i + "    "
-            if not tree.tail or not tree.tail.strip():
-                tree.tail = i
-            for tree in tree:
-                XMLFile.indent(tree, level + 1)
-            if not tree.tail or not tree.tail.strip():
-                tree.tail = i
-        else:
-            if level and (not tree.tail or not tree.tail.strip()):
-                tree.tail = i
-
-    def write_to_tar(self, tar, timestamp):
-        tarinfo = tarfile.TarInfo(self.filepath)
-        root = self.tree.getroot()
-        self.indent(root)
-        content = etree.tostring(root, encoding="utf-8")
-        tarinfo.size = len(content)
-        tarinfo.mtime = timestamp
-        tar.addfile(tarinfo, io.BytesIO(content))
-
 
 def update_git_repository():
     try:
@@ -102,106 +71,10 @@ def update_git_repository():
 def fetch_xml_files():
     os.chdir(root + "lensfun-git/data/db")
     xml_filenames = glob.glob("*.xml")
-    xml_files = set(XMLFile(os.getcwd(), filename) for filename in xml_filenames)
+    xml_files = set(XMLFile(filename, os.getcwd()) for filename in xml_filenames)
     timestamp = int(subprocess.check_output(["git", "log", "-1", '--format=%ad', "--date=raw", "--"] + xml_filenames). \
                     decode("utf-8").split()[0])
     return xml_files, timestamp
-
-
-class Converter:
-    from_version = None
-    to_version = None
-    def __call__(self, tree):
-        root = tree.getroot()
-        if self.to_version == 0:
-            if "version" in root.attrib:
-                del root.attrib["version"]
-        else:
-            root.attrib["version"] = str(self.to_version)
-
-converters = []
-current_version = 0
-def converter(converter_class):
-    global current_version
-    current_version = max(current_version, converter_class.from_version)
-    converters.append(converter_class())
-    return converter_class
-
-
-@converter
-class From1To0(Converter):
-    from_version = 1
-    to_version = 0
-
-    @staticmethod
-    def round_aps_c_cropfactor(lens_or_camera):
-        element = lens_or_camera.find("cropfactor")
-        if element is not None:
-            cropfactor = float(element.text)
-            if 1.5 < cropfactor < 1.56:
-                element.text = "1.5"
-            elif 1.6 < cropfactor < 1.63:
-                element.text = "1.6"
-
-    def __call__(self, tree):
-        super().__call__(tree)
-        for lens in tree.findall("lens"):
-            element = lens.find("aspect-ratio")
-            if element is not None:
-                lens.remove(element)
-            calibration = lens.find("calibration")
-            if calibration is not None:
-                for real_focal_length in calibration.findall("real-focal-length"):
-                    # Note that while one could convert it to the old
-                    # <field-of-view> element, we simply remove it.  It is not
-                    # worth the effort.
-                    calibration.remove(real_focal_length)
-            self.round_aps_c_cropfactor(lens)
-        for camera in tree.findall("camera"):
-            self.round_aps_c_cropfactor(camera)
-
-
-@converter
-class From2To1(Converter):
-    from_version = 2
-    to_version = 1
-
-    def __call__(self, tree):
-        super().__call__(tree)
-        for acm_model in tree.findall("//calibration/*[@model='acm']"):
-            acm_model.getparent().remove(acm_model)
-        for distortion in tree.findall("//calibration/distortion[@real-focal]"):
-            etree.SubElement(distortion.getparent(), "real-focal-length", {"focal": distortion.get("focal"),
-                                                                           "real-focal": distortion.get("real-focal")})
-            del distortion.attrib["real-focal"]
-
-
-def generate_database_tarballs(xml_files, timestamp):
-    version = current_version
-    output_path = os.path.join(args.output_path, "db")
-    shutil.rmtree(output_path, ignore_errors=True)
-    os.makedirs(output_path)
-    metadata = [timestamp, [], []]
-    while True:
-        metadata[1].insert(0, version)
-
-        tar = tarfile.open(os.path.join(output_path, "version_{}.tar.bz2".format(version)), "w:bz2")
-        for xml_file in xml_files:
-            xml_file.write_to_tar(tar, timestamp)
-        tar.close()
-
-        try:
-            converter_instance = converters.pop()
-        except IndexError:
-            break
-        assert converter_instance.from_version == version
-        for xml_file in xml_files:
-            converter_instance(xml_file.tree)
-        version = converter_instance.to_version
-    json.dump(metadata, open(os.path.join(output_path, "versions.json"), "w"))
-    if args.upload:
-        subprocess.check_call(["rsync", "-a", "--delete", output_path if output_path.endswith("/") else output_path + "/",
-                               config["SourceForge"]["login"] + "@web.sourceforge.net:/home/project-web/lensfun/htdocs/db"])
 
 
 def send_email(to, subject, body):
@@ -325,7 +198,13 @@ def close_github_issues():
 db_was_updated = update_git_repository()
 if db_was_updated:
     xml_files, timestamp = fetch_xml_files()
-    generate_database_tarballs(xml_files, timestamp)
+    output_path = os.path.join(args.output_path, "db")
+    xml_converter.generate_database_tarballs(xml_files, timestamp, output_path)
+    if args.upload:
+        subprocess.check_call(
+            ["rsync", "-a", "--delete", output_path if output_path.endswith("/") else output_path + "/",
+             config["SourceForge"]["login"] + "@web.sourceforge.net:/home/project-web/lensfun/htdocs/db"])
+
 owncloud.sync()
 close_github_issues()
 owncloud.sync()
